% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportnonparametrics.R
\docType{package}
\name{reportnonparametrics}
\alias{reportnonparametrics}
\title{reportnonparametrics Package}
\description{
Automatic report of findings from various nonparametric tests
Description: This function automatically creates reports of different normality tests.
    These tests are: Pearson's chi-squared test, Kruskal-Wallis rank sum test, Friedman rank sum test and Fisher's exact test
}
\seealso{
Relevant functions of this package:
- {report_nonparametrics()}
- {cite()}
}
\author{
Anna Huelemeier \email{anna@huelemeier.de}
}
